Rename main module to `elliptic_curves`.
