+++ elliptic_curves/elliptic/errors.py
