+++ elliptic_curves/polynomial/irreducible.py
@@ -1,6 +1,6 @@
-from src.parser.polynomial import parse_polynomial
-from src.polynomial.errors import UnknownIrreduciblePolynomialPower
-from src.polynomial.polynomial import Polynomial
+from elliptic_curves.parser.polynomial import parse_polynomial
+from elliptic_curves.polynomial.errors import UnknownIrreduciblePolynomialPower
+from elliptic_curves.polynomial.polynomial import Polynomial
 
 GF2_IRREDUCIBLE_POLYNOMIALS_DIRECTORY = {
     2: 'x^2+x+1',
