+++ elliptic_curves/polynomial/utils.py
