+++ elliptic_curves/settings.py
