+++ tests/polynomial/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
-from src.polynomial.utils import convert_bits_array_to_num
-from src.polynomial.utils import convert_num_to_bits_array
+from elliptic_curves.polynomial.utils import convert_bits_array_to_num
+from elliptic_curves.polynomial.utils import convert_num_to_bits_array
 
 
 @pytest.mark.parametrize(
