+++ tests/test_elliptic.py
@@ -1,12 +1,12 @@
 import pytest
 
-from src.elliptic.elliptic import GF2NotSupersingularCurve
-from src.elliptic.elliptic import GF2SupersingularCurve
-from src.elliptic.elliptic import Point
-from src.elliptic.elliptic import ZpCurve
-from src.polynomial.polynomial import Polynomial
-from src.polynomial.polynomial import polyone
-from src.polynomial.polynomial import polyzero
+from elliptic_curves.elliptic.elliptic import GF2NotSupersingularCurve
+from elliptic_curves.elliptic.elliptic import GF2SupersingularCurve
+from elliptic_curves.elliptic.elliptic import Point
+from elliptic_curves.elliptic.elliptic import ZpCurve
+from elliptic_curves.polynomial.polynomial import Polynomial
+from elliptic_curves.polynomial.polynomial import polyone
+from elliptic_curves.polynomial.polynomial import polyzero
 
 
 def pone() -> Polynomial:
